fix #3351: allow keeping a different number of snapshots on source and destination
by introducing a new dest-maxsnap parameter which can be used to override
maxsnap for the destination side.

This is useful for backups, as one can potentially save a lot of space on the
source side (or the destination side if one can come up with a use case for
that) by keeping fewer snapshots around.

Signed-off-by: Fabian Ebner <f.ebner@proxmox.com>