close #1824: add 'compressed' option enabling -c and -L for send
If one of the features lz4_compress, zstd_compress or large_blocks is
actively used/required by the source dataset, but not enabled on the
destination, zfs send/recv will fail and recommend upgrading the pool.

It's not required that the target dataset is also compressed.
Specifying '--properties' to have the target inherit the compression
property even works during later syncs (but pre-existing data won't be
recompressed automatically then).

Signed-off-by: Fabian Ebner <f.ebner@proxmox.com>